EmptyStateManager class

EmptyStateManager class

EmptyStateManager class

Methods

Public methods

EmptyStateManager$new()

EmptyStateManager$is_empty_state_show()

EmptyStateManager$show()

EmptyStateManager$hide()

EmptyStateManager$clone()

Method new()

Usage

EmptyStateManager$new(
  id,
  html_content = default_empty_state_component(),
  color = NULL
)

Arguments

id

id

html_content

html_content

color

color Returns the current visibility state of the empty state UI Show empty state Hide empty state

Method is_empty_state_show()

Usage

EmptyStateManager$is_empty_state_show()

Method show()

Usage

EmptyStateManager$show()

Method hide()

Usage

EmptyStateManager$hide()

Method clone()

The objects of this class are cloneable with this method.

Usage

EmptyStateManager$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.